96252: Move Bitstream page components out of SharedModule
